Move remaining previously-`App` tests to `command_test.go`
and remove duplicates
